# batch/tassidicambio/src/main/resources/openapi.yaml
openapi: 3.0.3
info:
  title: Swagger Exchange Rate BDI - OpenAPI 3.0
  description: >-
    Swagger Exchange Rate BDI
  version: 1.0.11
...
servers:
  - url: http://localhost:8080/api/v3
...
paths:
...
  /historyExchangeRates:
    post:
      summary: Visualizza, rispetto alla valuta e all'intervallo temporale inserite in request, i tassi di cambio della divisa passata in input.
      description: ''
      operationId: drawHistoryExchangeRates
      requestBody:
        description: ''
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ExchangeRatesRequest'
          application/xml:
            schema:
              $ref: '#/components/schemas/ExchangeRatesRequest'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/ExchangeRatesRequest'
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ExchangeRate'
            application/xml:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ExchangeRate'
        default:
          description: Default error sample response
      tags:
        - Exchange Rates
...
components:
  schemas:
...
    SortData:
      type: object
      properties:
        field:
          type: string
          example: avgRate
        sort:
          type: string
          enum:
            - ASC
            - DESC
          example: ASC
    ExchangeRatesRequest:
      type: object
      properties:
        from:
          type: string
          format: date
        to:
          type: string
          format: date
        referenceCurrency:
          type: string
          example: EUR
        isoCurrencies:
          type: array
          xml:
            name: currencies
            wrapped: true
          items:
            type: string
          example: ["USD", "AUD", "GBP"]
        page:
          type: integer
          format: int64
          example: 0
        nrElement:
          type: integer
          format: int64
          example: 20
        sort:
          $ref: '#/components/schemas/SortData'
      xml:
        name: exchangeRatesRequest
    ExchangeRate:
      type: object
      properties:
        country:
          type: string
          example: AUSTRALIA
        currency:
          type: string
          example: Dollaro Australiano
        isoCode:
          type: string
          example: AUD
        uicCode:
          type: integer
          format: int64
          example: 109
        avgRate:
          type: number
          format: float
          example: 1.6515
        exchangeConvention:
          type: string
          example: Quantità di valuta estera per 1 Euro
        referenceDate:
          type: string
          format: date
          example: 2024-02-26
        referenceCurrency:
          type: string
          example: EUR
        tmsInsert:
          type: string
          format: date-time
      xml:
        name: exchangeRate
...